Fix location of TestReaderWriter and DoubleReader helper projects
