+++ test/Shared/DoubleReader/DoubleReader1.cs
