+++ test/Shared/DoubleReader/DoubleReader2.cs
