+++ test/Shared/DoubleReader/Properties/AssemblyInfo.cs
