+++ test/Shared/TestReaderWriter/Properties/AssemblyInfo.cs
